requirements.txt
pandas
numpy
scikit-learn
statsmodels
understatapi
pyyaml
sqlalchemy
aiohttp
fastapi
uvicorn
python-dotenv
pydantic
certifi
